requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.10.6
numpy<2.0.0

# Audio processing
faster-whisper==1.0.3
pyannote.audio==3.1.1
torch==2.6.0
torchaudio==2.6.0

# LLM integration (HTTP OpenAI-compatible)
langchain==0.3.20
langchain-openai==0.3.8
langchain-core==0.3.43
httpx==0.28.1

# Utilities
tqdm==4.67.1